+++ source/logfiles.c
@@ -1,4 +1,4 @@
-/* $EPIC: logfiles.c,v 1.29 2005/01/13 16:06:06 jnelson Exp $ */
+/* $EPIC: logfiles.c,v 1.30 2005/03/11 05:02:22 jnelson Exp $ */
 /*
  * logfiles.c - General purpose log files
  *
@@ -683,7 +683,7 @@ BUILT_IN_COMMAND(logcmd)
 }
 
 /****************************************************************************/
-void	add_to_logs (int winref, int servref, const char *target, int level, const char *orig_str)
+void	add_to_logs (long winref, int servref, const char *target, int level, const char *orig_str)
 {
 	Logfile *log;
 	int	i;
